+++ threedi_schema/domain/models.py
@@ -8,22 +8,31 @@ Base = declarative_base()  # automap_base()
 
 
 class Lateral2D(Base):
-    __tablename__ = "v2_2d_lateral"
+    __tablename__ = "lateral_2d"
     id = Column(Integer, primary_key=True)
-
-    type = Column(IntegerEnum(constants.Later2dType), nullable=False)
-    the_geom = Column(Geometry("POINT"), nullable=False)
-    timeseries = Column(Text, nullable=False)
+    code = Column(Text)
+    display_name = Column(Text)
+    type = Column(IntegerEnum(constants.Later2dType))
+    timeseries = Column(Text)
+    time_units = Column(Text)
+    interpolate = Column(Boolean)
+    offset = Column(Integer)
+    units = Column(Text)
+    tags = Column(Text)
+    geom = Column(Geometry("POINT"), nullable=False)
 
 
 class BoundaryConditions2D(Base):
-    __tablename__ = "v2_2d_boundary_conditions"
+    __tablename__ = "boundary_condition_2d"
     id = Column(Integer, primary_key=True)
-
-    display_name = Column(String(255))
-    timeseries = Column(Text, nullable=False)
-    boundary_type = Column(IntegerEnum(constants.BoundaryType), nullable=False)
-    the_geom = Column(Geometry("LINESTRING"), nullable=False)
+    code = Column(Text)
+    display_name = Column(Text)
+    type = Column(IntegerEnum(constants.BoundaryType))
+    timeseries = Column(Text)
+    time_units = Column(Text)
+    interpolate = Column(Boolean)
+    tags = Column(Text)
+    geom = Column(Geometry("LINESTRING"), nullable=False)
 
 
 class ControlMeasureLocation(Base):
@@ -232,23 +241,24 @@ class GroundWater(Base):
         return self.groundwater_hydraulic_conductivity_file
 
 
-class GridRefinement(Base):
-    __tablename__ = "v2_grid_refinement"
+class GridRefinementLine(Base):
+    __tablename__ = "grid_refinement_line"
     id = Column(Integer, primary_key=True)
-
     display_name = Column(String(255))
-    refinement_level = Column(Integer, nullable=False)
-    the_geom = Column(Geometry("LINESTRING"), nullable=False)
+    grid_level = Column(Integer)
+    geom = Column(Geometry("LINESTRING"), nullable=False)
     code = Column(String(100))
+    tags = Column(Text)
 
 
 class GridRefinementArea(Base):
-    __tablename__ = "v2_grid_refinement_area"
+    __tablename__ = "grid_refinement_area"
     id = Column(Integer, primary_key=True)
     display_name = Column(String(255))
-    refinement_level = Column(Integer, nullable=False)
+    grid_level = Column(Integer)
     code = Column(String(100))
-    the_geom = Column(Geometry("POLYGON"), nullable=False)
+    geom = Column(Geometry("POLYGON"), nullable=False)
+    tags = Column(Text)
 
 
 class CrossSectionDefinition(Base):
@@ -274,20 +284,22 @@ class ConnectionNode(Base):
     code = Column(String(100))
 
     manholes = relationship("Manhole", back_populates="connection_node")
-    boundary_conditions = relationship(
-        "BoundaryCondition1D", back_populates="connection_node"
-    )
-    laterals1d = relationship("Lateral1d", back_populates="connection_node")
 
 
 class Lateral1d(Base):
-    __tablename__ = "v2_1d_lateral"
+    __tablename__ = "lateral_1d"
     id = Column(Integer, primary_key=True)
-    connection_node_id = Column(
-        Integer, ForeignKey(ConnectionNode.__tablename__ + ".id"), nullable=False
-    )
-    timeseries = Column(Text, nullable=False)
-    connection_node = relationship(ConnectionNode, back_populates="laterals1d")
+    code = Column(Text)
+    display_name = Column(Text)
+    timeseries = Column(Text)
+    time_units = Column(Text)
+    interpolate = Column(Boolean)
+    offset = Column(Integer)
+    units = Column(Text)
+    tags = Column(Text)
+    geom = Column(Geometry("POINT"), nullable=False)
+
+    connection_node_id = Column(Integer)
 
 
 class Manhole(Base):
@@ -439,7 +451,7 @@ class AggregationSettings(Base):
 class PhysicalSettings(Base):
     __tablename__ = "physical_settings"
     id = Column(Integer, primary_key=True)
-    use_advection_1d = Column(IntegerEnum(constants.OffOrStandard))
+    use_advection_1d = Column(IntegerEnum(constants.AdvectionTypes1D))
     use_advection_2d = Column(IntegerEnum(constants.OffOrStandard))
 
 
@@ -462,23 +474,19 @@ class TimeStepSettings(Base):
 
 
 class BoundaryCondition1D(Base):
-    __tablename__ = "v2_1d_boundary_conditions"
+    __tablename__ = "boundary_condition_1d"
 
     id = Column(Integer, primary_key=True)
-    boundary_type = Column(IntegerEnum(constants.BoundaryType), nullable=False)
-    timeseries = Column(Text, nullable=False)
+    code = Column(Text)
+    display_name = Column(Text)
+    type = Column(IntegerEnum(constants.BoundaryType))
+    timeseries = Column(Text)
+    time_units = Column(Text)
+    interpolate = Column(Boolean)
+    tags = Column(Text)
+    geom = Column(Geometry("POINT"), nullable=False)
 
-    connection_node_id = Column(
-        Integer,
-        ForeignKey(ConnectionNode.__tablename__ + ".id"),
-        nullable=False,
-        unique=True,
-    )
-    connection_node = relationship(
-        ConnectionNode,
-        foreign_keys=connection_node_id,
-        back_populates="boundary_conditions",
-    )
+    connection_node_id = Column(Integer)
 
 
 class SurfaceMap(Base):
@@ -520,8 +528,6 @@ class Channel(Base):
     cross_section_locations = relationship(
         "CrossSectionLocation", back_populates="channel"
     )
-    potential_breaches = relationship("PotentialBreach", back_populates="channel")
-    exchange_lines = relationship("ExchangeLine", back_populates="channel")
     exchange_thickness = Column(Float)
     hydraulic_conductivity_in = Column(Float)
     hydraulic_conductivity_out = Column(Float)
@@ -650,9 +656,12 @@ class Culvert(Base):
 
 
 class DemAverageArea(Base):
-    __tablename__ = "v2_dem_average_area"
+    __tablename__ = "dem_average_area"
     id = Column(Integer, primary_key=True)
-    the_geom = Column(Geometry("POLYGON"), nullable=False)
+    geom = Column(Geometry("POLYGON"), nullable=False)
+    display_name = Column(Text)
+    code = Column(Text)
+    tags = Column(Text)
 
 
 class Weir(Base):
@@ -754,41 +763,37 @@ class Pumpstation(Base):
 
 
 class Obstacle(Base):
-    __tablename__ = "v2_obstacle"
+    __tablename__ = "obstacle"
     id = Column(Integer, primary_key=True)
     code = Column(String(100))
-    crest_level = Column(Float, nullable=False)
-    the_geom = Column(Geometry("LINESTRING"), nullable=False)
+    crest_level = Column(Float)
+    geom = Column(Geometry("LINESTRING"), nullable=False)
+    tags = Column(Text)
+    display_name = Column(String(255))
 
 
 class PotentialBreach(Base):
-    __tablename__ = "v2_potential_breach"
+    __tablename__ = "potential_breach"
     id = Column(Integer, primary_key=True)
     code = Column(String(100))
     display_name = Column(String(255))
-    exchange_level = Column(Float)
-    maximum_breach_depth = Column(Float)
+    tags = Column(Text)
+    initial_exchange_level = Column(Float)
+    final_exchange_level = Column(Float)
     levee_material = Column(IntegerEnum(constants.Material))
-    the_geom = Column(Geometry("LINESTRING"), nullable=False)
-    channel_id = Column(
-        Integer, ForeignKey(Channel.__tablename__ + ".id"), nullable=False
-    )
-    channel = relationship(
-        Channel, foreign_keys=channel_id, back_populates="potential_breaches"
-    )
+    geom = Column(Geometry("LINESTRING"), nullable=False)
+    channel_id = Column(Integer)
 
 
 class ExchangeLine(Base):
-    __tablename__ = "v2_exchange_line"
+    __tablename__ = "exchange_line"
     id = Column(Integer, primary_key=True)
-    the_geom = Column(Geometry("LINESTRING"), nullable=False)
-    channel_id = Column(
-        Integer, ForeignKey(Channel.__tablename__ + ".id"), nullable=False
-    )
-    channel = relationship(
-        Channel, foreign_keys=channel_id, back_populates="exchange_lines"
-    )
+    geom = Column(Geometry("LINESTRING"), nullable=False)
+    channel_id = Column(Integer)
     exchange_level = Column(Float)
+    display_name = Column(Text)
+    code = Column(Text)
+    tags = Column(Text)
 
 
 class Tags(Base):
@@ -817,7 +822,7 @@ DECLARED_MODELS = [
     ExchangeLine,
     Floodfill,
     ModelSettings,
-    GridRefinement,
+    GridRefinementLine,
     GridRefinementArea,
     GroundWater,
     Interflow,